+++ app/routes/components/Tag/demos/Demo01.tsx
@@ -1,0 +1,5 @@
+import { Tag } from "src/components/Tag";
+
+export default function Demo() {
+  return <Tag>Copy UI</Tag>;
+}