+++ src/components/Tag/index.tsx
@@ -1,3 +1,3 @@
-// Tag from copy-ui @ 2025-09-26.
+// Tag from copy-ui @ 2025-10-21.
 
 export { Tag } from "./Tag";